+++ packages/core/src/CatalogPanel.stories.tsx
@@ -109,11 +109,11 @@ export const WithExample: ComponentStory<typeof CatalogPanel> = () => {
   return <CatalogPanel />
 }
 
-export const WithNodeLabel: ComponentStory<typeof CatalogPanel> = () => {
+export const WithNodeLegend: ComponentStory<typeof CatalogPanel> = () => {
   const setCatalog = useSetCatalog()
   setCatalog({
     title: 'Some title',
-    nodeLabel: 'My custom node label',
+    nodeLegend: 'My custom node label',
     categories: [],
     global: {
       schema: {},